Feature/room db

+++ core/data/datasource/remote/src/main/kotlin/com/demo/datausage/core/data/datasource/remote/dto/DataUsageDtos.kt
@@ -18,7 +18,7 @@ data class DataUsageResult(
 
 data class DataUsageRecord(
     @SerializedName("volume_of_mobile_data")
-    val volume:BigDecimal,
+    val volume:Double,
     @SerializedName("quarter")
     val quarter:String,
     @SerializedName("_id")
